requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.2.0
certifi==2019.3.9
colorama==0.4.1
faced==0.1
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
isort==4.3.9
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
mock==2.0.0
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
opencv-python==4.0.0.21
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.1
pylint==2.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.1
Werkzeug==0.15.2
wincertstore==0.2
wrapt==1.11.1
